refactor: (#56) features의 main 슬라이스 배럴 파일 생성 및 import 경로 수정
