+++ src/features/main/index.ts
@@ -1,0 +1,10 @@
+export * from './constants/filters';
+export * from './constants/roomSettings';
+export * from './constants/titles';
+export * from './lib/validators';
+export { useCreateRoomModalStore } from './model/useCreateRoomModalStore';
+export { useSearchStore } from './model/useSearchStore';
+export { CreateRoomModal } from './ui/CreateRoomModal/CreateRoomModal';
+export { MainHeader } from './ui/MainHeader/MainHeader';
+export { RoomList } from './ui/RoomList/RoomList';
+export { SideBar } from './ui/SideBar';
