+++ src/features/main/ui/CreateRoomModal/MaxPlayerSelector.tsx
@@ -1,4 +1,5 @@
 import { ChevronLeftIcon, ChevronRightIcon } from '@heroicons/react/24/outline';
+
 import { cn } from '@/shared/lib/cn';
 
 interface MaxPlayerSelectorProps {
