--- src/features/main/ui/index.ts
@@ -1,4 +1,0 @@
-export * from './RoomList';
-export * from './MainHeader';
-export * from './CreateRoomModal';
-export * from './SideBar';
